Issue #28: Move of YangSimpleCompletionProcessor to help package
